Enable hashing and comparison on self.rows
Now rows are comparable between script runs. If rows have not changed
since the last run, there is no need to inspect the table for edits.

In order to accomplish this, the _rows() method in Identifier returns
a tuple, which is hashable, and the _edit() method in Edit no longer
modifies its rows attribute.