Fix order and compostion of init variables
This was done to avoid a "Maximum Recursion Depth Exceeded" error,
since uncaught attributes were being funnelled to arcpy.Describe through
the __getattr__ method.

The other change was to remove a method that figured out the database
based on name length, since this is already assigned in the config file.