Fix bug with ID carrying over incorrectly
If an ID had an incorrect prefix but a valid ID (e.g. a culvert with
ID='SWP12'), The prefix would get QC'd, but the ID would not
(e.g. new_id='SWCU12'). This left a bug where the new ID potentially
already existed. Now if the prefix is incorrect, the ID automatically
gets QC'd.